+++ tests/Integration/Psalm/Fixture/ConsumerFacade.php
@@ -6,10 +6,13 @@ namespace GacelaTest\Integration\Psalm\Fixture;
 
 use Gacela\Framework\AbstractFacade;
 
+/**
+ * @extends AbstractFacade<ConsumerFactory>
+ */
 final class ConsumerFacade extends AbstractFacade
 {
     public function knownMethod(): string
     {
-        return 'known';
+        return $this->getFactory()->createKnown();
     }
 }